To_do/pyproject.toml
[project]
name = "to-do"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt==4.0.1",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
